-Bugfix: RealSignalOut put now overwrites both getSignalId()-methods. -Bugfix: Signal::getSignalById does now work correct.

+++ includes/eeros/types.hpp
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef uint32_t sigdim_t;
+typedef uint16_t sigdim_t;
 typedef uint32_t sigid_t;
 typedef uint16_t sigmajorid_t;
 typedef uint16_t sigindex_t;
